[IDLE-000] Center모듈 파일디렉토리 정리
