+++ project/Projects/Presentation/Feature/Center/Sources/Screen/RegisterPost/View/ApplicationDetailView.swift
